requirements.txt
Django==1.11.10
gevent==1.3.5
greenlet==0.4.14
pytz==2018.5
PyYAML==3.13
uWSGI==2.0.17.1
